xive: Change block numbering
Instead of making the block ID equal to the chip ID, we allocate block
numbers in a linear way. This allows us to pack block numbers which will
in turn give us a way to pick an aligned power of two set of block numbers
for use by block group mode.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>