xive: Fix xive_get_ir_targetting()
There are two issues, first we didn't return a prio 0xff for a masked
IVE, second we shouldn't fail if the EQ isn't found as this can be called
for an interrupt that has never been routed to a valid EQ yet. In that
case, we pick a reasonable default server and a prio of 0xff.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>