hdat: make HDIF_child() print more useful errors
The HDAT has very fixed relationships between a parent and child
structures. When parsing the HDAT we always know what type of child
that we expect and we will error out if there is a mismatch between what
we get and what expect. This patch just prints out what we got in
addition to what we expected to help with debugging.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>