hdat: stop using proc_int_line in favor of pir
This field has been deprecated in the P9 HDAT format. Due to how XICS
works proc_int_line has always been equal to the pir field of the same
structure. This equality has always been enforced by an assert() in skiboot
on P7 and P8 so it's safe to just look at the pir field instead.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>