hdat: add radix-AP-encodings
The the P9 Radix TLB management instructions use an arbitrary "AP" field
that specifies the actual page size. The values used in the AP field are
processor dependent and needs to be passed in via the devicetree by
firmware. This patch adds the property for P9 based systems that boot
via the HDAT path.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>