hdat: parse and export STOP levels
ISAv3 added the STOP instruction to act as a front end to the various
power saving modes. The exact function of the stop instruction is
implementation dependent and the firmware must advertise which "stop
levels" are supported.

This patch adds the required HDAT parsing and exports the
"ibm,enabled-stop-levels" devicetree node for use inside of skiboot.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>