xive: Fix configuration of the EX special BAR
This BAR is used to force a core to send "local" commands when hitting
the TM area which has the same address on all chips.

There is one copy of this per core *pair*. The existing code only sets
it up on core 0 of a pair. However some systems have core 1 and not
core 0 ... oooops.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>