lpc/uart: Support routing of selected LPC interrupts to Linux
Each LPC interrupt can be routed to one of 4 lines to the PSI
bridge which represent 4 different system interrupts. This
allows LPC clients to request as specific target (Linux or OPAL)
and makes the LPC core pick a route and configure it appropriately.

The UART is updated to properly forward interrupts to Linux
if necessary

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>