hdata/test: strip blobs from the DT output
Adds a post-processing step the hdata_to_dt test program that replaces
large "blobs" in the device tree with a <0xcafebeef, length, checksum>
u32 tuple.

The actual contents of these blobs are mostly irrelevant. We only care
that they are a) present in the device tree and b) aren't silently
corrupted by the HDAT parser. Outputting a checksum instead of the full
blob is sufficent for testing the parser.

This reduces the size of the test case output considerably so hopefully
we stop crashing everyone's mail client whenever the tests are changed.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>