astbmc/p8dnu: Enable PCI slot's power supply on PEX8718 in hot-add path
This issue is reported from superMicro's "p8dnu" platform. PEX8718
is connected to PHB direct slot. We create dynamic PCI slots for
its (2) downstream ports and all of them support surprise hotplug
capability. The problem is power supply lost on hot-remove and it
isn't turned on automatically on hot-add. It means the PCIe link
behind the slot isn't up and the PCI adapter behind the slot can't
be probed successfully.

This fixes the issue by forcing to turn on the power supply on
hardware when user (kernel) requests to do so, as we did for PEX9733.

Reported-by: Hank Chang <hankmax0000@gmail.com>
Signed-off-by: Gavin Shan <gwhsan@linux.vnet.ibm.com>
Tested-by: Willie Liauw <williel@supermicro.com.tw>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>