pci: Add a framework for quirks
In future we may want to be able to do fixups for specific PCI devices in
skiboot, so add a small framework for doing this.

This is not intended for the same purposes as quirks in the Linux kernel,
as the PCI devices that quirks can match for in skiboot are not properly
configured.  This is intended to enable having a custom path to make
changes that don't directly interact with the PCI device, for example
adding device tree entries.

Signed-off-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Reviewed-by: Andrew Donnellan <andrew.donnellan@au1.ibm.com>
Reviewed-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
[stewart@linux.vnet.ibm.com: fix 0 vs NULL sparse warning, (C) date]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>