core/pci: Fix lost NVMe adapter behind PMC 8546 switch
The NVMe adapter in below PCI topology is lost. The root cause is
the presence bit on its PCI slot is missed, but the PCIe link has
been up. The PCI core doesn't probe the adapter behind the slot,
leading to lost NVMe adapter in the particular case.

   PHB3 root port
   PLX switch 8748 (10b5:8748)
   PLX swich 9733 (10b5:9733)
   PMC 8546 swtich (11f8:8546)
   NVMe adapter (1c58:0023)

This fixes the issue by overriding the PCI slot presence bit with
PCIe link state bit.

Reported-by: Mark E Schreiter <markes@us.ibm.com>
Signed-off-by: Gavin Shan <gwshan@linux.vnet.ibm.com>
Tested-by: Mark E Schreiter <markes@us.ibm.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>