core/init: Use '_' as separator in names of "exports" properties
The names of the properties under /ibm,opal/firmware/exports are used
directly by Linux to create files in sysfs. To remain consistent with
the existing naming of OPAL sysfs files, use '_' as the separator.

In particular for the symbol map which is already exported separately,
it's cleaner for the two files to have the same name, eg:

  /sys/firmware/opal/exports/symbol_map
  /sys/firmware/opal/symbol_map

Fixes: 9ffbfe269ec6 ("core/init: Add hdat-map property to OPAL node")
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>