xive: Fix potential for lost IPIs when manipulating CPPR
We don't trigger the IPI in set_mfrr() if the CPPR of the
destination is more favored than the MFRR. However, we fail
to re-evaluate it when the CPPR later changes.

This fixes it. The way this is implemented can lead to
spurious IPIs but these are harmless.

(Mikey remove bogus line setting xs->ipi_sent)

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Michael Neuling <mikey@neuling.org>