mambo: Add skiboot/linux symbol lookup
Adds the skisym and linsym commands which can be used to find the
address of a Linux or Skiboot symbol. To function this requires
the user to provide the SKIBOOT_MAP and VMLINUX_MAP environmental
variables which indicate which skiboot.map and System.map files
should be used.

Examples:

Look up a symbol address:

	systemsim % skisym .load_and_boot_kernel
	0x0000000030013a08

Set a breakpoint there:

	systemsim % b [skisym .load_and_boot_kernel]
	breakpoint set at [0:0]: 0x0000000030013a08 (0x0000000030013A08) Enc:0x7D800026 : mfcr    r12

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>