ipmi-sel: Fix missing typo, line length and correct eSEL stylisation
It might seem impossible to some that missing typos can be fixed
without being first added, but here we are. Further, reduce the line
length to sensible limits to ensure Joel provides a Reviewed-by, and
fix eSEL stylisation to hopefully extract similar tags from the rest of
the peanut gallery.

Suggested-by: Joel Stanley <joel@jms.id.au>
Suggested-by: Russell Currey <ruscur@russell.cc>
Signed-off-by: Andrew Jeffery <andrew@aj.id.au>
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>